Remove the one line installation
If I use the one line installation as soon as I run `zokrates setup` it tells me I have a problem with malleability. At this point I'm a clueless noob, I have absolutely no idea what malleability is, except that it sounds bad. When I try to find more information, it is not trivial to understand.

You should either update the one line installation to include `libsnark`, or not have it this early in the process, IMAO.